8.1-Route_optimization_agent_crewAI/requirements.txt
crewai>=0.28.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
requests>=2.31.0 